refactor ONNXModel into own file
